+++ HISTORY.md
@@ -9,6 +9,7 @@ History of SixtyPical
     in a `copy` to a vector is a routine that is defined further down in the source.
 *   Initialized `word` memory locations.
 *   Can `copy` a literal word to a word table.
+*   Subtract word (constant or memory location) from word memory location.
 *   Fixed bug which was preventing `if` branches to diverge in what they initialized,
     if it was already initialized when going into the `if`.
 *   Fixed a bug which was making it crash when trying to analyze `repeat forever` loops.